+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased] - yyyy-mm-dd
+
+### Changed
+- Removed siridb password from get settings base handler output
+
 ## [0.1.0-beta3.2.3] - 2022-05-02
 
 ### Added